fix floating point division in level formatter
(cherry picked from commit a50e3e392de6a30fa5e59a9589a035c83f90ee53)

+++ src/main/java/ru/zznty/create_factory_logistics/logistics/panel/FactoryFluidPanelBehaviour.java
@@ -77,9 +77,9 @@ public class FactoryFluidPanelBehaviour extends FactoryPanelBehaviour implements
                 .color(satisfied ? 0xD7FFA8 : promisedSatisfied ? 0xffcd75 : 0xFFBFA8)
                 .add(CreateLang.text(promised == 0 ? "" : "\u23F6"))
                 .add(CreateLang.text("/")
-                        .style(ChatFormatting.WHITE))
+                             .style(ChatFormatting.WHITE))
                 .add(formatLevel(count)
-                        .color(0xF1EFE8))
+                             .color(0xF1EFE8))
                 .component();
     }
 
@@ -96,12 +96,12 @@ public class FactoryFluidPanelBehaviour extends FactoryPanelBehaviour implements
     public ValueSettingsBoard createBoard(Player player, BlockHitResult hitResult) {
         int maxAmount = 100;
         return new ValueSettingsBoard(CreateLang.translate("factory_panel.target_amount")
-                .component(), maxAmount, 10,
-                List.of(CreateLang.translate("schedule.condition.threshold.buckets")
-                                .component(),
-                        CreateLang.translate("schedule.condition.threshold.buckets")
-                                .component()),
-                new ValueSettingsFormatter(this::formatValue));
+                                              .component(), maxAmount, 10,
+                                      List.of(CreateLang.translate("schedule.condition.threshold.buckets")
+                                                      .component(),
+                                              CreateLang.translate("schedule.condition.threshold.buckets")
+                                                      .component()),
+                                      new ValueSettingsFormatter(this::formatValue));
     }
 
     public static LangBuilder formatLevel(int level) {
@@ -115,7 +115,8 @@ public class FactoryFluidPanelBehaviour extends FactoryPanelBehaviour implements
         if (level < 100)
             return CreateLang.number(level).add(CreateLang.translate("generic.unit.millibuckets"));
 
-        return CreateLang.number(Math.round((float) level / 100) / 10.).add(CreateLang.translate("generic.unit.buckets"));
+        return CreateLang.number(Math.round((float) level / 100) / 10.).add(
+                CreateLang.translate("generic.unit.buckets"));
     }
 
     public static LangBuilder formatLevel(int level, boolean round) {
@@ -124,16 +125,16 @@ public class FactoryFluidPanelBehaviour extends FactoryPanelBehaviour implements
 
         if (!round || level < 1000 || level % 1000 != 0) {
             if (level % 1000 == 0)
-                return CreateLang.number(level / 1000).add(CreateLang.translate("generic.unit.buckets"));
+                return CreateLang.number((float) level / 1000).add(CreateLang.translate("generic.unit.buckets"));
 
             return CreateLang.number(level).add(CreateLang.translate("generic.unit.millibuckets"));
         }
 
         if (level < 100_001) {
-            return CreateLang.number(level / 1000).add(CreateLang.translate("generic.unit.buckets"));
+            return CreateLang.number((float) level / 1000).add(CreateLang.translate("generic.unit.buckets"));
         }
 
-        return CreateLang.number(level / 1_000_000)
+        return CreateLang.number((float) level / 1_000_000)
                 .add(CreateLang.text("k"))
                 .add(CreateLang.translate("generic.unit.buckets"));
     }
